requirements.txt
discord.py>=2.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
